requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Django==3.2.9
djangorestframework==3.12.4
faiss-cpu==1.7.1.post3
filelock==3.8.0
huggingface-hub==0.8.1
idna==3.3
joblib==1.1.0
nltk==3.7
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
transformers==4.21.1
typing_extensions==4.3.0
urllib3==1.26.11
